Move RefreshWebcalJob logic to a proper service so that it may be called independently
